Refactor : OtherVideo 컴포넌트 분리, OtherVideo에서 ref와 class 이용해 마이크, 캠 상태 표시 하던 것을 수정 OtherVideo는 이제 Props로 넘겨 받은 member의 mic, cam으로 mute표시와 camOff 표시를 조건부 렌더링 합니다
